Add Tensor[List]View::element_type. Move element_type trait to tensor_view.h. Move element_t to util.h. Other minor fixes (max_errors, some style).
